+++ front/index.html
@@ -8,6 +8,7 @@
   <link rel="icon" type="image/x-icon" href="/icons/iconv1.png">
   
   <!-- Fonts -->
+  <link rel="preconnect" href="https://fonts.gstatic.com" />
   <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
   <style>
     @import url('https://fonts.googleapis.com/css2?family=Grandstander:ital,wght@0,100..900;1,100..900&display=swap');
