+++ front/src/components/App/Animals/AnimalList.tsx
@@ -111,7 +111,8 @@ function AnimalList() {
       <div className="md:my-3 flex flex-wrap font-body w-full bg-zoning rounded-lg shadow dark:bg-gray-800 justify-around">
         <form className="text-texte justify-around" onSubmit={handleSubmit}>
           <div id="fullSearch" className="mx-2 col-span-3 items-center flex flex-wrap justify-around">
-            <h3 className="font-grands text-2xl w-full my-2 text-center">Rechercher un animal</h3>
+            <h2 className="font-grands text-2xl w-full my-2 text-center">Rechercher un animal</h2>
+            <label htmlFor="espece-dropdown-small">Par espèce</label>
             <select onChange={handleInputData("especeDropdownSmall")} tabIndex={0} className="col-span-3 text-xs block w-[50%]" id="espece-dropdown-small" name="especeDropdownSmall" defaultValue="defaultSmall">
               <option value="defaultSmall" disabled hidden>--Choisissez une espèce--</option>
               {speciesItems}
