+++ front/src/components/App/Header/HeaderLogin.tsx
@@ -7,19 +7,21 @@ function HeaderLogin() {
             <Link to="/connexion" className="hover:text-accents1-light item-link">Se Connecter</Link>
           </li>
           
-          <div id="log-in" className="flex flex-row place-self-center md:place-self-start pl-2 gap-2">
-            <li className="border-2 border-accents2-dark place-self-center md:place-self-start pl-2 item-link">
-              <p>S'Inscrire :</p>
-            </li>
+          <li>
+            <ul id="log-in" className="flex flex-row place-self-center md:place-self-start pl-2 gap-2">
+              <li className="border-2 border-accents2-dark place-self-center md:place-self-start pl-2 item-link">
+                <p>S'Inscrire :</p>
+              </li>
 
-            <li className="border-2 border-accents2-dark">
-              <Link to="/famille/inscription" className="hover:text-accents1-light item-link">Famille</Link>
-            </li>
-            
-            <li className="border-2 border-accents2-dark">
-              <Link to="/association/inscription" className="hover:text-accents1-light item-link">Association</Link>
-            </li>
-          </div>
+              <li className="border-2 border-accents2-dark">
+                <Link to="/famille/inscription" className="hover:text-accents1-light item-link">Famille</Link>
+              </li>
+              
+              <li className="border-2 border-accents2-dark">
+                <Link to="/association/inscription" className="hover:text-accents1-light item-link">Association</Link>
+              </li>
+            </ul>
+          </li>
     </>
   )
 }
