+++ front/src/components/App/Shelter/ShelterRequestList.tsx
@@ -1,6 +1,7 @@
 import { Link } from "react-router-dom";
 import { useUserContext } from "../../../contexts/UserContext";
 import { useRootContext } from '../../../contexts/RootContext';
+import ShelterRequestTable from "./ShelterRequestTable";
 
 function ShelterRequestList() {
   const { animals } = useRootContext();
@@ -16,38 +17,8 @@ function ShelterRequestList() {
   const requested = animals.filter(({ association_id, demandes }) => Number(association_id) === Number(user.id) && demandes.length);
 
   const requestedAnimals = requested.map((animal) => (
-    <>
-    <tr onClick={handleClick} key={`${animal.id} header`} tabIndex={0} className="view text-fond text-sm bg-accents2 font-grands font-semibold p-3 border-accents2-dark border-solid border-1 hover:bg-accents2-dark">
-      <td colSpan={4} scope="colgroup" className="px-2 pt-2 border-accents2-dark border-solid border-1">{animal.nom}</td>
-      <td colSpan={2} scope="colgroup" className="px-2 pt-2 border-accents2-dark border-solid border-1">Nombre</td>
-    </tr>
-      <tr key={`${animal.id} body title`} className="fold text-fond text-sm bg-accents2-light font-grands font-semibold p-3 border-accents2-dark border-solid border-1 hidden">
-        <td colSpan={2} className="px-2 pt-2  border-accents2-light border-solid border-1">Famille</td>
-        <td colSpan={2} className="px-2 pt-2  border-accents2-light border-solid border-1">Date de demande</td>
-        <td colSpan={2} className="px-2 pt-2  border-accents2-light border-solid border-1">Statut</td>
-      </tr>
-      { animal.demandes.map((demande : any, index : any) => (
-      <tr key={`${animal.id} request n° ${demande.id}`} className={"fold text-sm font-body font-semibold hidden " + (index % 2 > 0 ? 'text-fond bg-accents2-light' : 'bg bg-fond')}>
-        
-          <td colSpan={2}>{demande.nom}</td>
-          <td colSpan={2}>{demande.Demande.date_debut}</td>
-          <td colSpan={2}><Link tabIndex={0} className="hover:underline" to={`/associations/profil/demandes/${demande.Demande.id}`} >{demande.Demande.statut_demande}</Link></td> 
-      </tr>
-      ))}
-    </>
+    <ShelterRequestTable key={animal.id} animal={animal} />
   ));
-
-  function handleClick(e: any) {
-    const fold = e.currentTarget.nextSibling;
-    fold.classList.toggle('hidden')
-  
-    let content = fold.nextSibling;
-
-    while(content && !content.classList.contains('font-grands')) {
-      content.classList.toggle('hidden');
-      content = content.nextSibling;
-    }
-   };
 
   return(
     <main className="justify-self-stretch flex-1">
@@ -80,10 +51,12 @@ function ShelterRequestList() {
                 <>
                 <h4 className="w-full text-center font-grands text-2xl my-4">Demandes en cours</h4>
                 <table className="table text-center w-full md:w-5/6">
-                  <tr key={"header"} className="border-none bg-zoning text-sm font-grands">
-                    <td colSpan={4} scope="colgroup">Nom Animal</td>
-                    <td colSpan={2} scope="colgroup">Nombre de demandes</td>
-                  </tr>
+                  <thead className="border-none bg-zoning text-sm font-grands">
+                    <tr>
+                      <th colSpan={4} scope="colgroup">Nom Animal</th>
+                      <th colSpan={2} scope="colgroup">Nombre de demandes</th>
+                    </tr>
+                  </thead>
                   {requestedAnimals}
                 </table>
                 </>
